requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Data manipulation and analysis
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Progress bars and utilities
tqdm>=4.65.0

# Optional: for faster data loading
pillow>=10.0.0

# For confusion matrix
scikit-learn>=1.3.0

# API and Web Server
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
kagglehub>=0.3.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
